+++ src/file/file.c
@@ -26,7 +26,8 @@
 #include <stdbool.h>
 
 linted_error linted_file_create(linted_ko * restrict kop, linted_ko dirko,
-                                char const *pathname, int flags, mode_t mode)
+                                char const *pathname, unsigned long flags,
+                                mode_t mode)
 {
     linted_error errnum;
